cloud_functions/analyze_file/requirements.txt
google-cloud-firestore==1.6.1
google-cloud-language==1.3.0
google-cloud-storage==1.25.0
google-cloud-vision==0.41.0
